fix(TranslateService): Reverted the use of Injector and changed Angular 2 requirement to ~beta.0 ins
It was not a good idea to use the Injector to load Http, the real problem was that Ionic 2 uses

Angular 2 beta.0 and ng2-translate uses Angular 2 beta.1. The ionic 2 cli was packaging 2 different

instances of Angular 2 and ng2-translate couldn't get Http because the provider it used wasn't the

same as the one from the main Ionic 2 application. Changing the requirement to ~beta.0 fixes that

problem.