feat(TranslationService): getting translations for default lang when it's set
Fixes #332
Fixes #336

BREAKING CHANGE: we will now fetch the translations for the default lang when it is set (if they are not already available). Also the default lang is now automatically set to the first lang that receives translations (if it is still undefined at that time). Setting the default lang will override this default behavior, so this will probably not break anything for you, but if you set the translations for the default lang first then you don't even have to call `setDefaultLang` anymore.